+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+- drop IP address family requirement from firewall rule
+
 ## [4.5.2]
 
 ### Fixed